fixed cart spawn
cant click them anymore, just dragging now
you get your coins back when placed outside of grid

+++ Assets/XR/Esmée/Scripts/Minions/MinionSO/MinionScriptableObject.cs
@@ -5,7 +5,7 @@ public class MinionScriptableObject : ScriptableObject
 {
     public float MSpeed;
     public float MDamage;
-    public float MCost;
+    public int MCost;
     public int MLevel;
     public int MPrize;
     public GameObject MPrefab;
